requirements.txt
flask
Flask-SQLAlchemy
flask-loginFlask-WTF==1.2.2
